move FeatureDescriptor to polymer dir
