+++ src/ast/ast.ts
@@ -16,7 +16,6 @@ export * from './descriptor';
 export * from './document-descriptor';
 export * from './element-descriptor';
 export * from './event-descriptor';
-export * from './feature-descriptor';
 export * from './import-descriptor';
 export * from './inline-document-descriptor';
 export * from './property-descriptor';
